requirements.txt
Flask
gunicorn
pymysql
Authlib
python-dotenv